# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crypto Bot API",
        "description": "API para gerenciamento de portfólios de criptomoedas, estratégias de trading e alertas",
        "contact": {
            "name": "Suporte Crypto Bot",
            "email": "support@cryptobot.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/exchanges": {
            "get": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Listar exchanges",
                "description": "Listar todas as exchanges do usuário autenticado.",
                "operationId": "b31e993848843cedcbc518d6ea3d9928",
                "parameters": [
                    {
                        "name": "portfolio_id",
                        "in": "query",
                        "description": "ID do portfólio para filtrar exchanges",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de exchanges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "portfolio_id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "spot",
                                                    "futures"
                                                ]
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            },
                                            "testnet": {
                                                "type": "boolean"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Criar exchange",
                "description": "Criar uma nova exchange para um portfólio do usuário autenticado.",
                "operationId": "c5b448efaa9cfbbd59038acb1c956354",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "portfolio_id",
                                    "name",
                                    "api_key",
                                    "api_secret",
                                    "type"
                                ],
                                "properties": {
                                    "portfolio_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "api_key": {
                                        "type": "string"
                                    },
                                    "api_secret": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "spot",
                                            "futures"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "testnet": {
                                        "type": "boolean"
                                    },
                                    "rate_limit": {
                                        "type": "integer"
                                    },
                                    "last_sync": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Exchange criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "portfolio_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/exchanges/{id}": {
            "get": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Exibir exchange",
                "description": "Exibir uma exchange específica do usuário autenticado.",
                "operationId": "7c669903355b77d117f54233924b9b99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da exchange",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da exchange",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "portfolio_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exchange não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Atualizar exchange",
                "description": "Atualizar uma exchange do usuário autenticado.",
                "operationId": "5eaaec19be66169c581dac81db50f979",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da exchange",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "api_key": {
                                        "type": "string"
                                    },
                                    "api_secret": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "spot",
                                            "futures"
                                        ]
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "testnet": {
                                        "type": "boolean"
                                    },
                                    "rate_limit": {
                                        "type": "integer"
                                    },
                                    "last_sync": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Exchange atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "portfolio_id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "type": {
                                            "type": "string"
                                        },
                                        "is_active": {
                                            "type": "boolean"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exchange não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Exchanges"
                ],
                "summary": "Remover exchange",
                "description": "Remover uma exchange do usuário autenticado.",
                "operationId": "0715960fc1a1c4846f05ccbc85623386",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da exchange",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exchange removida com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Exchange não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}